# docs/swagger.yaml
basePath: /
definitions:
  dto.AuthResponse:
    properties:
      access_token:
        example: eyJhbGciOiJIUzI1NiIs...
        type: string
      expires_in:
        example: 900
        type: integer
      refresh_token:
        example: eyJhbGciOiJIUzI1NiIs...
        type: string
      user:
        $ref: '#/definitions/dto.UserResponse'
    type: object
  dto.LoginDTO:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  dto.MessageResponse:
    properties:
      message:
        example: logged out
        type: string
    type: object
  dto.RefreshTokenDTO:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  dto.RegisterDTO:
    properties:
      email:
        type: string
      password:
        maxLength: 72
        minLength: 8
        type: string
    required:
    - email
    - password
    type: object
  dto.TokenPairResponse:
    properties:
      access_token:
        example: eyJhbGciOiJIUzI1NiIs...
        type: string
      expires_in:
        example: 900
        type: integer
      refresh_token:
        example: eyJhbGciOiJIUzI1NiIs...
        type: string
    type: object
  dto.UserListResponse:
    properties:
      total:
        example: 42
        type: integer
      users:
        items:
          $ref: '#/definitions/dto.UserResponse'
        type: array
    type: object
  dto.UserResponse:
    properties:
      created_at:
        example: "2026-04-12T00:00:00Z"
        type: string
      email:
        example: user@example.com
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      role:
        example: user
        type: string
      status:
        example: active
        type: string
      updated_at:
        example: "2026-04-12T00:00:00Z"
        type: string
    type: object
  httpserver.Response:
    properties:
      data: {}
      error: {}
      request_id:
        type: string
      trace_id:
        type: string
    type: object
host: localhost
info:
  contact: {}
  description: Centralized Exchange demo — token trading platform
  title: Booker CEX API
  version: "1.0"
paths:
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Login request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.LoginDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httpserver.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.AuthResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/httpserver.Response'
            - properties:
                error:
                  type: object
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/httpserver.Response'
            - properties:
                error:
                  type: object
              type: object
      summary: Login with email and password
      tags:
      - auth
  /api/v1/auth/logout:
    post:
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httpserver.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.MessageResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/httpserver.Response'
            - properties:
                error:
                  type: object
              type: object
      security:
      - BearerAuth: []
      summary: Logout (revoke all tokens)
      tags:
      - auth
  /api/v1/auth/me:
    get:
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httpserver.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.UserResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/httpserver.Response'
            - properties:
                error:
                  type: object
              type: object
      security:
      - BearerAuth: []
      summary: Get current authenticated user
      tags:
      - auth
  /api/v1/auth/refresh:
    post:
      consumes:
      - application/json
      parameters:
      - description: Refresh token request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.RefreshTokenDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httpserver.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.TokenPairResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/httpserver.Response'
            - properties:
                error:
                  type: object
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/httpserver.Response'
            - properties:
                error:
                  type: object
              type: object
      summary: Refresh access token
      tags:
      - auth
  /api/v1/auth/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: Register request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.RegisterDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/httpserver.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.AuthResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/httpserver.Response'
            - properties:
                error:
                  type: object
              type: object
        "409":
          description: Conflict
          schema:
            allOf:
            - $ref: '#/definitions/httpserver.Response'
            - properties:
                error:
                  type: object
              type: object
      summary: Register a new user
      tags:
      - auth
  /api/v1/users:
    get:
      parameters:
      - default: 20
        description: Limit
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset
        in: query
        name: offset
        type: integer
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httpserver.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.UserListResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/httpserver.Response'
            - properties:
                error:
                  type: object
              type: object
      security:
      - BearerAuth: []
      summary: List users
      tags:
      - users
  /api/v1/users/{id}:
    get:
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httpserver.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.UserResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/httpserver.Response'
            - properties:
                error:
                  type: object
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/httpserver.Response'
            - properties:
                error:
                  type: object
              type: object
      security:
      - BearerAuth: []
      summary: Get user by ID
      tags:
      - users
securityDefinitions:
  BearerAuth:
    description: Enter "Bearer {token}"
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"